create basemodal and use it
